+++ src/components/CharacterSheet/CharacterSheet.scss
@@ -62,7 +62,7 @@
       .combat-values-wrapper {
         grid-area: armor;
       }
-      .inventory {
+      .weapons {
         grid-area: weapons;
       }
     }
