--- src/components/DisplayInventory/DisplayInventory.scss
@@ -1,173 +1,0 @@
-@import 'src/variables';
-
-.inventory {
-  padding: 2em;
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-
-  .weapons {
-    display: flex;
-    flex-direction: column;
-    gap: 1em;
-
-    .weapon {
-      display: grid;
-      grid-template-rows: 1fr 1fr 2fr 1fr 2fr;
-      grid-template-columns: 1fr 1fr 3fr 2fr 2fr;
-      grid-template-areas:
-        'title title title . .'
-        'title title title attack-bonus-title critical-hit-title'
-        'name-input name-input name-input attack-bonus-input critical-hit-input'
-        'type-title range-title ammunition-title damage-title damage-title'
-        'type-input range-input ammunition-input damage-input damage-input';
-
-      .title {
-        background-color: $blackEquivalent;
-        color: white;
-        padding: 0.2em 1em;
-        text-align: center;
-      }
-
-      .input {
-        padding: 0.1em 0.5em;
-        border-top-width: 0;
-        font-size: 1.5em;
-        width: 100%;
-      }
-
-      .header-container {
-        display: flex;
-        align-items: stretch;
-        border: none;
-        position: relative;
-
-        .header {
-          flex-grow: 1;
-          font-size: 1.5em;
-          border: none;
-
-          .delete-weapon {
-            cursor: pointer;
-            font-size: 0.9em;
-            position: absolute;
-            right: 3em;
-            border-radius: 100%;
-            border: transparent solid 1px;
-
-            &:hover {
-              border-color: $blackEquivalent;
-            }
-          }
-        }
-      }
-
-      .type-input {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-      }
-
-      .critical-hit-input {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-
-        input {
-          //font-size: inherit;
-          font-size: 1em;
-          width: 3em;
-        }
-      }
-
-      // table cell borders
-      $borderWidth: 1px;
-
-      & > * {
-        border: $borderWidth solid $blackEquivalent;
-      }
-
-      .type-input,
-      .range-input,
-      .ammunition-input,
-      .damage-input {
-        border-bottom-width: #{2 * $borderWidth};
-      }
-
-      .critical-hit-title,
-      .critical-hit-input,
-      .damage-title,
-      .damage-input {
-        border-right-width: #{2 * $borderWidth};
-      }
-
-      .name-input,
-      .type-title,
-      .type-input {
-        border-left-width: #{2 * $borderWidth};
-      }
-
-      .name-input {
-        border-top-color: transparent;
-      }
-
-      // grid areas
-      .header-container {
-        grid-area: title;
-      }
-
-      .name-input {
-        grid-area: name-input;
-      }
-
-      .attack-bonus-title {
-        grid-area: attack-bonus-title;
-      }
-
-      .attack-bonus-input {
-        grid-area: attack-bonus-input;
-      }
-
-      .critical-hit-title {
-        grid-area: critical-hit-title;
-      }
-
-      .critical-hit-input {
-        grid-area: critical-hit-input;
-      }
-
-      .type-title {
-        grid-area: type-title;
-      }
-
-      .type-input {
-        grid-area: type-input;
-      }
-
-      .range-title {
-        grid-area: range-title;
-      }
-
-      .range-input {
-        grid-area: range-input;
-      }
-
-      .ammunition-title {
-        grid-area: ammunition-title;
-      }
-
-      .ammunition-input {
-        grid-area: ammunition-input;
-      }
-
-      .damage-title {
-        grid-area: damage-title;
-      }
-
-      .damage-input {
-        grid-area: damage-input;
-      }
-    }
-  }
-}
